Forms 1 first experiments: - Role of smoothness. - Role of other preprocessing steps. - Attempts to locate forms (toy model of toy model of forms, sparse parity) - Initial changes to incorporate rLLC

+++ src/icl/regression/train.py
@@ -124,7 +124,8 @@ class RegressionRun:
             self.optimizer.load_state_dict(last_checkpoint["optimizer"])
 
             if "rng_state" in last_checkpoint:
-                torch.set_rng_state(torch.tensor(last_checkpoint["rng_state"]))
+                torch.set_rng_state(torch.tensor(last_checkpoint["rng_state"], dtype=torch.uint8))
+
 
             if self.scheduler is not None:
                 self.scheduler.load_state_dict(last_checkpoint["scheduler"])
